# docs/swagger.yaml
basePath: /
definitions:
  github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.CheckoutMessage:
    properties:
      status:
        type: string
      total:
        type: integer
    type: object
  github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.CheckoutRequest:
    properties:
      orderItems:
        items:
          $ref: '#/definitions/github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.OrderItemRequest'
        type: array
    type: object
  github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.CheckoutResponse:
    properties:
      code:
        type: string
      message:
        $ref: '#/definitions/github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.CheckoutMessage'
      order_id:
        type: integer
    type: object
  github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.DeleteUserRequest:
    properties:
      email:
        type: string
    type: object
  github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.LoginResponse:
    properties:
      auth:
        type: string
      code:
        type: string
    type: object
  github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.Offer:
    properties:
      category:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        type: string
      price:
        type: integer
      quantity:
        type: integer
      updatedAt:
        type: string
    type: object
  github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.OffersResponse:
    properties:
      code:
        type: string
      message:
        items:
          $ref: '#/definitions/github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.Offer'
        type: array
    type: object
  github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.OrderItemRequest:
    properties:
      productID:
        type: integer
      quantity:
        type: integer
    type: object
  github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.OrderStatusResponse:
    properties:
      code:
        type: string
      status:
        type: string
    type: object
  github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.OrderStatusUpdateRequest:
    properties:
      status:
        type: string
    type: object
  github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.OrderStatusUpdateResponse:
    properties:
      code:
        type: string
      status:
        type: string
    type: object
  github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.RegisterUserRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.Response:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.Response_d:
    properties:
      code:
        type: string
      message: {}
    type: object
  github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.User:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      id:
        type: integer
      password:
        type: string
      token:
        type: string
      updatedAt:
        type: string
      username:
        type: string
    type: object
  github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.UsersResponse:
    properties:
      code:
        type: string
      users:
        items:
          $ref: '#/definitions/github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.User'
        type: array
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
host: localhost:3000
info:
  contact:
    email: fiber@swagger.io
    name: API Support
  description: This is a sample Swagger for Fiber
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fiber Example API
  version: "1.0"
paths:
  /admin/dashboard:
    get:
      consumes:
      - application/json
      description: Get the admin dashboard
      parameters:
      - description: JWT <token>
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.Response_d'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.Response_d'
        "500":
          description: Bad server
          schema:
            $ref: '#/definitions/github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.Response_d'
      security:
      - ApiKeyAuth: []
      summary: Admin dashboard
      tags:
      - admin
  /admin/orders/{id}:
    patch:
      consumes:
      - application/json
      description: Update the status of a specific order by id. Valid statuses are
        "preparing", "processing", "shipped", "delivered".
      parameters:
      - description: JWT <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: Order Status Update Request
        in: body
        name: updateRequest
        required: true
        schema:
          $ref: '#/definitions/github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.OrderStatusUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: status
          schema:
            $ref: '#/definitions/github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.OrderStatusUpdateResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.Response'
        "500":
          description: Bad server
          schema:
            $ref: '#/definitions/github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.Response'
      security:
      - ApiKeyAuth: []
      summary: Update the status of a specific order
      tags:
      - admin
  /admin/users:
    delete:
      consumes:
      - application/json
      description: Remove a customer by email, only for admins
      parameters:
      - description: JWT <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Delete User Request
        in: body
        name: email
        required: true
        schema:
          $ref: '#/definitions/github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.DeleteUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.Response'
        "500":
          description: Bad server
          schema:
            $ref: '#/definitions/github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.Response'
      security:
      - ApiKeyAuth: []
      summary: Remove a customer
      tags:
      - admin
    get:
      consumes:
      - application/json
      description: Get all buyers, only for admins
      parameters:
      - description: JWT <token>
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: users
          schema:
            $ref: '#/definitions/github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.UsersResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.Response'
        "500":
          description: Bad server
          schema:
            $ref: '#/definitions/github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.Response'
      security:
      - ApiKeyAuth: []
      summary: Get all buyers
      tags:
      - admin
  /auth/checkout:
    post:
      consumes:
      - application/json
      description: 'Buy a list of orders. If you want to add more products, here is
        an example of the structure to follow:{"orderItems": [ { "productID": 1, "quantity":
        2 }, { "productID": 2, "quantity": 1 }, { "productID": 3, "quantity": 5 }
        ] }'
      parameters:
      - description: Checkout Request
        in: body
        name: checkout
        required: true
        schema:
          $ref: '#/definitions/github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.CheckoutRequest'
      - description: JWT <token>
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.CheckoutResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.Response'
        "500":
          description: Bad server
          schema:
            $ref: '#/definitions/github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.Response'
      security:
      - ApiKeyAuth: []
      summary: Checkout
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login a user with the given credentials
      parameters:
      - description: Login Request
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: token
          schema:
            $ref: '#/definitions/github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.LoginResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.Response'
        "500":
          description: Bad server
          schema:
            $ref: '#/definitions/github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.Response'
      summary: Login a user
      tags:
      - auth
  /auth/offers:
    get:
      consumes:
      - application/json
      description: Get all available offers
      parameters:
      - description: JWT <token>
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: offers
          schema:
            $ref: '#/definitions/github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.OffersResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.Response'
        "500":
          description: Bad server
          schema:
            $ref: '#/definitions/github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.Response'
      security:
      - ApiKeyAuth: []
      summary: Get available offers
      tags:
      - auth
  /auth/orders/{id}:
    get:
      consumes:
      - application/json
      description: Get the status of a specific order by id
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: JWT <token>
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: status
          schema:
            $ref: '#/definitions/github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.OrderStatusResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.Response'
        "500":
          description: Bad server
          schema:
            $ref: '#/definitions/github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.Response'
      security:
      - ApiKeyAuth: []
      summary: Get status of a specific order
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user with the given details
      parameters:
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.RegisterUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User added
          schema:
            $ref: '#/definitions/github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.Response'
        "500":
          description: Bad server
          schema:
            $ref: '#/definitions/github_com_ICOMP-UNC_newworld-gastonsegura2908_git_internal_models.Response'
      summary: Register a new user
      tags:
      - auth
swagger: "2.0"